Create new google-cloud-core gem package
This gem will be shared across all google-cloud gems.
