+++ google-cloud-core/Gemfile
@@ -1,0 +1,16 @@
+source "http://rubygems.org"
+
+gemspec
+
+# TODO: unpin rake after YARD (and other tasks) no longer depend on last_comment
+gem "rake", "~> 10.0"
+
+gem "gcloud-rdoc",
+    git: "https://github.com/GoogleCloudPlatform/gcloud-ruby.git",
+    branch: "gcloud-rdoc"
+gem "yard-gcloud",
+    git: "https://github.com/GoogleCloudPlatform/gcloud-ruby.git",
+    branch: "yard-gcloud"
+gem "gcloud-jsondoc",
+    git: "https://github.com/GoogleCloudPlatform/gcloud-ruby.git",
+    branch: "gcloud-jsondoc"
