+++ google-cloud-core/lib/cloud.rb
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 
-require "google/cloud/version"
+require "google/cloud/core/version"
 
 ##
 # # Google Cloud
