+++ google-cloud-core/lib/google/cloud/credentials.rb
