+++ google-cloud-core/lib/google/cloud/errors.rb
