+++ google-cloud-core/test/google/cloud/errors/gapi_errors_test.rb
