+++ google-cloud-core/test/google/cloud/errors/grpc_errors_test.rb
